requirements.txt
streamlit
beautifulsoup4
requests
openai
sentence-transformers
scikit-learn
python-dotenv